Move RPCBus and Event Bus into their own packages
