+++ pkg/p2p/wire/event.go
@@ -2,14 +2,9 @@ package wire
 
 import (
 	"bytes"
-	"io"
 
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire/topics"
-	lg "github.com/sirupsen/logrus"
 )
-
-// QuitTopic is the topic to make all components quit
-const QuitTopic = "quit"
 
 type (
 
@@ -62,53 +57,6 @@ type (
 		Collect(*bytes.Buffer) error
 	}
 
-	// TopicProcessor is the interface for preprocessing events belonging to a specific topic
-	TopicProcessor interface {
-		Process(*bytes.Buffer) (*bytes.Buffer, error)
-	}
-
-	// TopicListener accepts events from the EventBus and takes care of reacting on
-	// quit Events. It delegates the business logic to the EventCollector which is
-	// supposed to handle the incoming events
-	TopicListener struct {
-		subscriber     EventSubscriber
-		eventCollector EventCollector
-		msgChan        <-chan *bytes.Buffer
-		msgChanID      uint32
-		quitChan       chan *bytes.Buffer
-		quitChanID     uint32
-		topic          string
-	}
-
-	// EventPreprocessor allows registration of preprocessors to be applied to incoming Event on a specific topic
-	EventPreprocessor interface {
-		RegisterPreprocessor(string, ...TopicProcessor) []uint32
-		RemovePreprocessor(string, uint32)
-		RemoveAllPreprocessors(string)
-	}
-
-	// EventSubscriber subscribes a channel to Event notifications on a specific topic
-	EventSubscriber interface {
-		EventPreprocessor
-		Subscribe(string, chan<- *bytes.Buffer) uint32
-		SubscribeCallback(string, func(*bytes.Buffer) error) uint32
-		SubscribeStream(string, io.WriteCloser) uint32
-		Unsubscribe(string, uint32)
-		// RegisterPreprocessor(string, ...TopicProcessor)
-	}
-
-	// EventPublisher publishes serialized messages on a specific topic
-	EventPublisher interface {
-		Publish(string, *bytes.Buffer)
-		Stream(string, *bytes.Buffer)
-	}
-
-	// EventBroker is an EventPublisher and an EventSubscriber
-	EventBroker interface {
-		EventSubscriber
-		EventPublisher
-	}
-
 	// EventDeserializer is the interface for those struct that allows deserialization of an event from scratch
 	EventDeserializer interface {
 		Deserialize(*bytes.Buffer) (Event, error)
@@ -123,64 +71,6 @@ type (
 		All() []Event
 	}
 )
-
-// NewTopicListener creates the TopicListener listening to a topic on the EventBus.
-// The EventBus, EventCollector and Topic are injected
-func NewTopicListener(subscriber EventSubscriber, collector EventCollector, topic string,
-	preprocessors ...TopicProcessor) *TopicListener {
-
-	msgChan := make(chan *bytes.Buffer, 100)
-	quitChan := make(chan *bytes.Buffer, 1)
-	msgChanID := subscriber.Subscribe(topic, msgChan)
-	quitChanID := subscriber.Subscribe(string(QuitTopic), quitChan)
-
-	if len(preprocessors) > 0 {
-		subscriber.RegisterPreprocessor(topic, preprocessors...)
-	}
-
-	return &TopicListener{
-		subscriber:     subscriber,
-		msgChan:        msgChan,
-		msgChanID:      msgChanID,
-		quitChan:       quitChan,
-		quitChanID:     quitChanID,
-		topic:          topic,
-		eventCollector: collector,
-	}
-}
-
-func (n *TopicListener) log() *lg.Entry {
-	return lg.WithFields(lg.Fields{
-		"topic":   n.topic,
-		"process": "TopicListener",
-	})
-}
-
-// Accept incoming (mashalled) Events on the topic of interest and dispatch them to the EventCollector.Collect.
-func (n *TopicListener) Accept() {
-	log := n.log().WithField("id", n.msgChanID)
-	log.Debugln("Accepting messages")
-	for {
-		select {
-		case <-n.quitChan:
-			n.subscriber.Unsubscribe(n.topic, n.msgChanID)
-			n.subscriber.Unsubscribe(string(QuitTopic), n.quitChanID)
-			return
-		case eventBuffer := <-n.msgChan:
-			if len(n.msgChan) > 10 {
-				log.WithField("Unconsumed", len(n.msgChan)).Debugln("Channel is accumulating messages")
-			}
-			if err := n.eventCollector.Collect(eventBuffer); err != nil {
-				log.WithError(err).Errorln("Error in eventCollector.Collect")
-			}
-		}
-	}
-}
-
-// Quit will kill the goroutine spawned by Accept, and unsubscribe from it's subscribed topics.
-func (n *TopicListener) Quit() {
-	n.quitChan <- new(bytes.Buffer)
-}
 
 // AddTopic is a convenience function to add a specified topic at the start of
 // a buffer. This topic is later decoded by the peer when gossiping messages,
