+++ pkg/util/nativeutils/eventbus/handlermap.go
@@ -1,4 +1,4 @@
-package wire
+package eventbus
 
 import (
 	"math/rand"
